+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - In path tracing, camera ray misses now return a null value with Minimum Depth > 1.
 - HD's SpeedTree 8 upgrader now sets up CullModeForward as well.
 - Restructured data under Display Stats panel to use column layout.
+- Added controls for the users to manually feed the ray tracing acceleration structure that should be used for a given camera (case 1370678).
 
 ## [13.1.0] - 2021-09-24